backend/requirements.txt
#langchain used for Document processing and chunking,LLM chain creation,Vector store integration, Memory management for conversations
langchain==0.1.0
langchain-community==0.0.16
langchain-core==0.1.20

#DB Integrations, Connecting to PostgreSQL, Storing and querying vector embeddings, Performing similarity searches
psycopg2-binary==2.9.9
pgvector==0.2.4

# fast api is modern web ramework, uvicorn is ASGI server, Creating REST API endpoints, Serving the application
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6

# Data Validation, validating req/response data, token count,Managing Environment variable configuration
pydantic>=2.5.3
python-dotenv==1.0.0

# Document processing
unstructured==0.10.30
pypdf==3.17.1
python-docx==1.0.1
beautifulsoup4==4.12.2
numpy==1.26.0
pandas==2.1.1

# Ollama integration
ollama==0.1.6 